fix database minor exception

+++ app/src/main/java/de/westwing/campaignbrowser/model/Campaign.kt
@@ -4,13 +4,13 @@ import android.os.Parcel
 import android.os.Parcelable
 
 data class Campaign(
-    val id: Long = 0,
+    val id: Int = 0,
     val name: String,
     val description: String,
     val imageUrl: String = ""
 ) : Parcelable {
     constructor(parcel: Parcel) : this(
-        parcel.readLong(),
+        parcel.readInt(),
         parcel.readString()!!,
         parcel.readString()!!,
         parcel.readString()!!
@@ -18,7 +18,7 @@ data class Campaign(
     }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeLong(id)
+        parcel.writeInt(id)
         parcel.writeString(name)
         parcel.writeString(description)
         parcel.writeString(imageUrl)
